move all cursing to v2.1
