+++ chains/evm/deployment/v1_6_0/adapters/init.go
@@ -6,6 +6,7 @@ import (
 	_ "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/hooks" // registers EVM post-proposal CCIP send hook provider
 	adapters1_2_0 "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v1_2_0/adapters"
 	evmseq "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v1_6_0/sequences"
+	rmnadapters "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v2_1_0/adapters"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/deploy"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/fastcurse"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/fees"
@@ -18,12 +19,15 @@ func init() {
 	v1_6_0 := utils.Version_1_6_0
 	v1_6_3 := utils.Version_1_6_3
 
+	// Cursing a 1.6 lane suite goes through the RMN 2.1.0 adapter. That adapter resolves the RMN via
+	// RMNProxy.getARM() rather than a pinned datastore ref, so it follows whichever RMN the chain
+	// actually runs -- which, after the RMN 2.1.0 rollout, is 2.1.0 even on a 1.6 lane suite.
 	curseRegistry := fastcurse.GetCurseRegistry()
 	curseRegistry.RegisterNewCurse(fastcurse.CurseRegistryInput{
 		CursingFamily:       chainsel.FamilyEVM,
 		CursingVersion:      v1_6_0,
-		CurseAdapter:        NewCurseAdapter(),
-		CurseSubjectAdapter: NewCurseAdapter(),
+		CurseAdapter:        rmnadapters.NewCurseAdapter(),
+		CurseSubjectAdapter: rmnadapters.NewCurseAdapter(),
 	})
 
 	laneMigratorRegistry := deploy.GetLaneMigratorRegistry()
